cloud-functions/drive-webhook/requirements.txt
functions-framework==3.0.0
google-cloud-storage==2.7.0
google-api-python-client==2.86.0
google-auth==2.17.3
requests==2.28.1
flask==2.0.1
werkzeug==2.0.1